fix : scale of everything is now tied to _gridSize (NO MORE ZINDEX ISSUE)

+++ gui/include/graphics/raylib/PlayerRaylib.hpp
@@ -16,7 +16,7 @@ namespace Zappy {
     class PlayerRaylib : public AEntityRaylib {
     public:
         PlayerRaylib(const std::shared_ptr<Player> &worldPlayer, PokemonInfo &pkInfo,
-            size_t gridSize, raylib::Shader &shader);
+            float gridSize, raylib::Shader &shader);
         void draw(const raylib::Camera camera, bool selectionMode);
         void update(void);
         void loadTextureAndModel(raylib::Shader &shader);
